Add basic test for the time skew problem.

+++ src/backend/access/transam/csn_snapshot.c
@@ -383,7 +383,7 @@ GenerateCSN(bool locked)
 	 * TODO: create some macro that add small random shift to current time.
 	 */
 	INSTR_TIME_SET_CURRENT(current_time);
-	csn = (SnapshotCSN) INSTR_TIME_GET_NANOSEC(current_time) + csn_time_shift*1E9;
+	csn = (SnapshotCSN) INSTR_TIME_GET_NANOSEC(current_time) + (uint64) (csn_time_shift * 1E9);
 
 	/* TODO: change to atomics? */
 	if (!locked)
